requirements.txt
opencv-python==4.9.*
pytesseract
py-sudoku
